requirements.txt
# LLM and Agent Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20

# Data Processing
requests==2.31.0
pyjstat==2.4.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.11.0

# Financial Data and Analysis
yfinance==0.2.25
pandas_datareader==0.10.0
alpha_vantage==2.3.1
quantlib==1.30
pyfolio==0.9.2

# Time Series and Forecasting
statsmodels==0.14.0
prophet==1.1

# Natural Language Processing (NLP)
spacy==3.5.0
transformers==4.30.0

# Utilities
python-dotenv==1.0.0
pytest==7.4.3

# Scientific Libraries
numpy==1.26.2
scipy==1.11.4
